add science department and research points producing

+++ Assets/Scripts/Data/PlayerData.cs
@@ -6,12 +6,12 @@ using UniRx;
 public class PlayerData
 {
     public ReactiveProperty<float> playerCredits { get; private set; }
-    public ReactiveProperty<int> researchPoints { get; private set; }
+    public ReactiveProperty<float> researchPoints { get; private set; }
 
     public PlayerData()
     {
         playerCredits = new ReactiveProperty<float>();
-        researchPoints = new ReactiveProperty<int>();
+        researchPoints = new ReactiveProperty<float>();
     }
 
     public Dictionary<string, object> ToDictionary()
@@ -29,7 +29,7 @@ public class PlayerData
         if (dict.TryGetValue("playerCredits", out object credits))
             playerData.playerCredits.Value = Convert.ToSingle(credits);
         if (dict.TryGetValue("researchPoints", out object points))
-            playerData.researchPoints.Value = Convert.ToInt32(points);
+            playerData.researchPoints.Value = Convert.ToSingle(points);
         return playerData;
     }
 }
